+++ src/main/java/com/blockchain/timebank/dao/ViewRequestDetailDao.java
@@ -10,9 +10,9 @@ import java.sql.Timestamp;
 import java.util.List;
 
 public interface ViewRequestDetailDao extends CrudRepository<ViewRequestDetailEntity, Long> {
-    List<ViewRequestDetailEntity> findViewRequestDetailEntitiesByServiceTypeAndIsDeletedOrderByCreateTimeDesc(String type, Byte deleted);
+    List<ViewRequestDetailEntity> findViewRequestDetailEntitiesByServiceTypeAndIsDeletedOrderByCreateTimeDesc(String type, int deleted);
     ViewRequestDetailEntity findViewRequestDetailEntityById(long id);
-    List<ViewRequestDetailEntity> findViewRequestDetailEntitiesByUserIdAndIsDeleted(Long id, byte isDeleted);
+    List<ViewRequestDetailEntity> findViewRequestDetailEntitiesByUserIdAndIsDeleted(Long id, int isDeleted);
     @Query("select e from ViewRequestDetailEntity e where e.serviceType=?1 and e.price <= ?2 and e.price >=?3 and e.endTime <= ?4 and e.endTime>= ?5 and e.beginTime <= ?4 and e.beginTime >= ?5 and e.serviceName in ?6")
     List<ViewRequestDetailEntity> findViewRequestDetailEntityByConditionOrderByCreateTimeDesc(String type, BigDecimal upperPrice, BigDecimal lowerPrice, Timestamp upperTime, Timestamp lowerTime, String[] serviceNameArr);
